Coding for Means, Some attempt at Likert Graphs
I've finished the coding for means of interest, knowledge, and employablity of Theology and Religious Studies by the other subjects. I've installed the required packages for making a graph of the Likert responses (categorical to go along with the already created means ones), but have not been able to figure how it works yet. I'll have another look at it later and work with it more.